AO2 from live

+++ wp-content/plugins/delete-me/inc/delete_user.php
@@ -11,6 +11,21 @@ if ( empty( $this->user_ID ) || $this->GET[$this->info['trigger']] != $this->use
 // Nonce
 if ( isset( $this->GET[$this->info['nonce']] ) == false || wp_verify_nonce( $this->GET[$this->info['nonce']], $this->info['nonce'] ) == false ) return; // stop executing file
 
+// User - Used to confirm password and for data in email notification
+$user = get_user_by( 'id', $this->user_ID );
+
+// Confirm Password Required?
+$your_profile_confirmation_page = ( is_admin() && basename( $_SERVER['SCRIPT_NAME'] ) == 'options.php' && $this->GET['page'] == ( $this->info['slug_prefix'] . '_confirmation' ) ) ? true : false;
+if ( ( $your_profile_confirmation_page && $this->option['settings']['your_profile_confirm_password_required'] ) || ( !$your_profile_confirmation_page && $this->option['settings']['shortcode_form_enabled'] ) ) {
+	
+	$this->POST = $this->striptrim_deep( $_POST );
+	$password_entered = '';
+	$password_entered = isset( $this->POST[$this->info['trigger'] . '_your_profile_confirm_password'] ) ? $this->POST[$this->info['trigger'] . '_your_profile_confirm_password'] : $password_entered;
+	$password_entered = isset( $this->POST[$this->info['trigger'] . '_shortcode_password'] ) ? $this->POST[$this->info['trigger'] . '_shortcode_password'] : $password_entered;
+	if ( empty( $password_entered ) || empty( $user ) || wp_check_password( $password_entered, $user->data->user_pass ) !== true ) return; // stop executing file
+	
+}
+
 // Include required WordPress function files
 include_once( ABSPATH . WPINC . '/post.php' ); // wp_delete_post
 include_once( ABSPATH . 'wp-admin/includes/bookmark.php' ); // wp_delete_link
@@ -23,6 +38,12 @@ if ( is_multisite() ) {
 	include_once( ABSPATH . 'wp-admin/includes/ms.php' ); // wpmu_delete_user
 	
 }
+
+// Number of Sites to which the user is registered
+$num_blogs_of_user = is_multisite() ? count( get_blogs_of_user( $this->user_ID ) ) : 1;
+
+// "Delete From Network" setting is checked or user belongs to only one blog
+$delete_from_network = ( is_multisite() && ( $this->option['settings']['ms_delete_from_network'] == true || $num_blogs_of_user == 1 ) ) ? true : false;
 
 // Posts
 //->>> Start: WordPress wp_delete_user Post types to delete
@@ -65,7 +86,7 @@ if ( $this->option['settings']['delete_comments'] == true ) :
 		
 		$comments_list[] = $comment['comment_ID'];
 		
-		// Delete comments if option set		
+		// Delete comments if option set
 		wp_delete_comment( $comment['comment_ID'] );
 		
 	}
@@ -77,28 +98,29 @@ if ( $this->option['settings']['email_notification'] == true ) :
 	
 	$email = array();
 	$email['to'] = get_option( 'admin_email' );
-	$email['subject'] = '[' . wp_specialchars_decode( get_option( 'blogname' ), ENT_QUOTES ) . '] Deleted User Notification';
+	$email['subject'] = sprintf( _x( '[%s] Deleted User Notification', '%s = site name', 'delete-me' ), wp_specialchars_decode( get_option( 'blogname' ), ENT_QUOTES ) );
 	$email['message'] =
-	'Deleted user on your site ' . wp_specialchars_decode( get_option( 'blogname' ), ENT_QUOTES ) . ':' . "\n\n" .
-	'Username: ' . $this->user_login . "\n\n" .	
-	'E-mail: ' . $this->user_email . "\n\n" .	
-	'This user deleted themselves using the WordPress plugin ' . $this->info['name'] . "\n\n" .	
-	count( $posts_list ) . ' Post(s)' . "\n" .
+	sprintf( _x( 'Deleted user on your site %s', '%s = site name', 'delete-me' ), wp_specialchars_decode( get_option( 'blogname' ), ENT_QUOTES ) ) . ":" . "\n\n" .
+	( ( $num_blogs_of_user > 1 && $delete_from_network ) ? ( sprintf( __( 'User deleted from %d total Network Sites.', 'delete-me' ), $num_blogs_of_user ) . "\n\n" ) : '' ) .
+	__( 'Username', 'delete-me' ) . ': ' . $this->user_login . "\n\n" .
+	__( 'E-mail', 'delete-me' ) . ': ' . $this->user_email . "\n\n" .
+	sprintf( _x( 'Registered: %s', '%s = date user registered with the Network or Site', 'delete-me' ), $user->data->user_registered ) . "\n\n" .
+	sprintf( _x( 'This user deleted themselves using the WordPress plugin %s', '%s = plugin name', 'delete-me' ), $this->info['name'] ) . "\n\n" .
+	sprintf( __( '%d Post(s)', 'delete-me' ), count( $posts_list ) ) . "\n" .
 	'----------------------------------------------------------------------' . "\n" .
-	implode( "\n\n", $posts_list ) . "\n\n" .	
-	count( $links_list ) . ' Link(s)' . "\n" .
+	implode( "\n\n", $posts_list ) . "\n\n" .
+	sprintf( __( '%d Link(s)', 'delete-me' ), count( $links_list ) ) . "\n" .
 	'----------------------------------------------------------------------' . "\n" .
-	implode( "\n\n", $links_list ) . "\n\n" .	
-	count( $comments_list ) . ' Comment(s)';	
+	implode( "\n\n", $links_list ) . "\n\n" .
+	sprintf( __( '%d Comment(s)', 'delete-me' ), count( $comments_list ) );
 	wp_mail( $email['to'], $email['subject'], $email['message'] );
 	
 endif;
 
 // Delete user
-if ( is_multisite() && $this->option['settings']['ms_delete_from_network'] == true && count( get_blogs_of_user( $this->user_ID ) ) == 1 ) {
+if ( $delete_from_network ) {
 	
 	// Multisite: Deletes user's Posts and Links, then deletes from WP Users|Usermeta
-	// ONLY IF "Delete From Network" setting checked and user only belongs to this blog	
 	wpmu_delete_user( $this->user_ID );
 	
 } else {
